Password Dumper
The .bat file aims at showing the SSID's resident on your computer or someones computer and dump them in a text file.

The aim however is not to just stop there l want to automate the whole thing do an autorun and when l plug in a flash then it does everything for me and l plug it out without even clicking the bat file.